pyproject.toml
[project]
name = "legal-smart-contracts-ontology"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "click>=8.1.8",
    "rdflib>=6.3.2",
    "slither-analyzer>=0.11.3",
]
